Remove binomial ll-inner term lookup table, and streamline tests to use 2D data input.
